reoder files / move setting styles to separate file / Settings
